+++ include/odroid/gpio.hpp
@@ -18,7 +18,7 @@ void sleep(std::chrono::milliseconds duration);
 template <typename Pin,
           typename... Args,
           typename std::enable_if_t<std::is_base_of_v<BasePin, std::decay_t<Pin>>, int> = 0>
-auto get(uint8_t pin_number, Args&&... args) -> Pin&
+auto get(uint16_t pin_number, Args&&... args) -> Pin&
 {
    return PinDatabase::get<Pin>(pin_number, args...);
 }
