+++ src/gpio.cpp
@@ -8,9 +8,9 @@ auto gpio::is_ready() -> bool
    return wiringPiSetup() != -1;
 }
 
-void gpio::pin_mode(uint8_t pin_number, uint8_t mode)
+void gpio::pin_mode(uint16_t pin_number, uint16_t mode)
 {
-   pinMode(pin_number, static_cast<int>(mode));
+   pinMode(static_cast<int>(pin_number), static_cast<int>(mode));
 }
 
 void gpio::sleep(std::chrono::milliseconds duration)
